[refactoring] norm수정
