+++ main.c
@@ -48,120 +48,6 @@ int	min_exit_num_check(char num, int i)
 	return (0);
 }
 
-void	line_token_quote(t_cmd *cmd, char *line, int *line_i, int *i)
-{
-	char	*tmp;
-
-	cmd->line_split[*line_i] = ms_line_tokenizing_quote(cmd, line, i);
-	while (ft_strchr(" |><", line[*i]) == NULL && line[*i])
-	{
-		if (line[*i] == DOUBLE_QUOTE || line[*i] == SINGLE_QUOTE)
-		{
-			tmp = ms_line_tokenizing_quote(cmd, line, i);
-			if (tmp == NULL)
-			{
-				cmd->line_split[++(*line_i)] = NULL;
-				i = 0;
-				while (cmd->line_split[*i])
-					free(cmd->line_split[(*i)++]);
-				free(cmd->line_split);
-				return ;
-			}
-		}
-		else
-			tmp = ms_line_tokenizing_str(cmd, line, i);
-		cmd->line_split[*line_i] = ft_strjoin(cmd->line_split[*line_i], tmp);
-	}
-	(*line_i)++;
-}
-
-static void	line_token_redirect_join(t_cmd *cmd, char *line, int *line_i, int *i)
-{
-	char	*tmp;
-
-	tmp = ms_line_tokenizing_str(cmd, line, i);
-	cmd->line_split[*line_i] = ft_strjoin(cmd->line_split[*line_i], tmp);
-	if (cmd->line_split[*line_i] == NULL)
-		exit (EXIT_FAILURE);
-	free(tmp);
-}
-
-void	line_token_redirect(t_cmd *cmd, char *line, int *line_i, int *i)
-{
-	if (line[*i] == '>')
-	{
-		cmd->line_split[*line_i] = ms_line_tokenizing_str(cmd, line, i);
-		if (line[*i] == '>')
-			line_token_redirect_join(cmd, line, line_i, i);
-	}
-	else
-	{
-		cmd->line_split[*line_i] = ms_line_tokenizing_str(cmd, line, i);
-		if (line[*i] == '<')
-			line_token_redirect_join(cmd, line, line_i, i);
-	}
-	(*line_i)++;
-}
-
-void	line_token_str(t_cmd *cmd, char *line, int *line_i, int *i)
-{
-	char	*tmp;
-
-	cmd->line_split[*line_i] = ms_line_tokenizing_str(cmd, line, i);
-	if (cmd->line_split[*line_i] == NULL)
-		exit (EXIT_FAILURE);
-	while (ft_strchr(" |><", line[*i]) == NULL && line[*i])
-	{
-		if (line[*i] == DOUBLE_QUOTE || line[*i] == SINGLE_QUOTE)
-		{
-			tmp = ms_line_tokenizing_quote(cmd, line, i);
-			if (tmp == NULL)
-			{
-				cmd->line_split[++(*line_i)] = NULL;
-				line_split_free(cmd);
-				return ;
-			}
-		}
-		else
-			tmp = ms_line_tokenizing_str(cmd, line, i);
-		cmd->line_split[*line_i] = ft_strjoin(cmd->line_split[*line_i], tmp);
-		free(tmp);
-	}
-	(*line_i)++;
-}
-
-void	ms_line_tokenizer(t_cmd *cmd, char *line)
-{
-	int		i = 0;
-	int		line_i = 0;
-
-	i = 0;
-	line_i = 0;
-	cmd->line_split = (char **)malloc(sizeof(char *) * (ft_strlen(line) + 1));
-	if (cmd->line_split == NULL)
-		exit (EXIT_FAILURE);
-	while (line[i])
-	{
-		if (line[i] == ' ')
-			while (line[i] == ' ')
-				i++;
-		else if (line[i] == DOUBLE_QUOTE || line[i] == SINGLE_QUOTE)
-			line_token_quote(cmd, line, &line_i, &i);
-		else if (line[i] == '|')
-			cmd->line_split[line_i++] = ms_line_tokenizing_str(cmd, line, &i);
-		else if (line[i] == '>' || line[i] == '<')
-			line_token_redirect(cmd, line, &line_i, &i);
-		else
-			line_token_str(cmd, line, &line_i, &i);
-	}
-	cmd->line_split[line_i] = NULL;
-	cmd->line_i = line_i;
-
-	i = 0;
-	while(cmd->line_split[i])
-		printf("check:%s\n", cmd->line_split[i++]);
-}
-
 void	ms_builtin_func(t_cmd *cmd)
 {
 	int		dir;
@@ -309,10 +195,13 @@ void	ms_builtin_func(t_cmd *cmd)
 
 void	ms_line_str_parsing(t_cmd *cmd)
 {
-	ms_line_tokenizer(cmd, cmd->line);	
+	ms_line_tokenizer(cmd, cmd->line);
+	int i = 0;
+	while(cmd->line_split[i])
+		printf("check:%s\n", cmd->line_split[i++]);
 	ms_builtin_func(cmd);
 
-	int	i = 0;
+	i = 0;
 	while (cmd->line_split[i])
 		free(cmd->line_split[i++]);
 	free(cmd->line_split);
