+++ ms_line_token_utils.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ms_line_tokenizing.c                               :+:      :+:    :+:   */
+/*   ms_line_token_utils.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jinsecho <jinsecho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/28 16:33:09 by jinsecho          #+#    #+#             */
-/*   Updated: 2024/10/28 16:35:53 by jinsecho         ###   ########.fr       */
+/*   Updated: 2024/10/28 20:17:35 by jinsecho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 static int	quote_flag_len(char *line, char c, int *i, int *flag)
 {
 	int	k;
-	
+
 	k = 0;
 	if (c == DOUBLE_QUOTE)
 		*flag = 1;
@@ -54,9 +54,9 @@ char	*ms_line_tokenizing_quote(t_cmd *cmd, char *line, int *i)
 
 char	*ms_line_tokenizing_str(t_cmd *cmd, char *line, int *i)
 {
-	char	*ptr;
 	int		k;
-	
+	char	*ptr;
+
 	k = 0;
 	ptr = (char *)malloc(sizeof(char) * ft_strlen(line) + 1);
 	if (ptr == NULL)
@@ -68,7 +68,7 @@ char	*ms_line_tokenizing_str(t_cmd *cmd, char *line, int *i)
 		while (ft_strchr(" |><\t\n", line[*i + k]) == NULL \
 		&& line[*i + k] != DOUBLE_QUOTE \
 		&& line[*i + k] != SINGLE_QUOTE && line[*i + k])
-		k++;
+			k++;
 	}
 	ft_strlcpy(ptr, line + *i, k + 1);
 	(*i) += k;
